Fixed data generator error on OSC
When I tried to make the data generators a class object they were convoluted and needed to be optimized. When I tried to run them they failed on OSC. I removed getting values and use a keras sequence generator to get the values for regions. I now just build a pool of genomic regions and use those regions + a random cell line to get the values.