whole-genome prediction error fix
There was an error for whole-genome prediction that caused it to crash after chromosome9 because it was not in the expected order of the header of the bigwig file. I fixed the chrom sizes file and sorted it by number correctly. I also removed the sorted() function that ordered the chromosomes incorrectly.