#24 Added np.nan_to_num to convert None to 0
When we import the data for prediction we are not replacing None values with 0. This is different than the approach that we used with bulk data analysis. The reason we want to do this is because we are import None values and I am not sure how that changes our predictions.