#35 Parser fix
Added some code to the train.py file to select epoch before exiting. Edited the name of a default chromosome list for training and validation to make it more clear. Changed the formatting of the output loss and dice figure. Dropped the tsv files for ROIs and also only write the combined files now for training and validation. Only report loss and dice now instead of spearman, pearson, precision, and recall for binary predictions.