#45 Added rev_comp flag, #47 Added pyyaml req, #46 fix invalid interval error
This push adds a flag to training that will allow you randomly generate the reverse complement examples for training in addition to the reference samples. The pyyaml req was added. The invalid interval bounds error was also patched by trying to get the input matrix and passing a matrix of zeros if the interval is not found.