optional modifications to parallel prediction
I have included some updated suggestions for parallel prediction. 1) Prediction starts with predict.py creating a pool of regions to predict on or formatting an input list of bed regions for prediction. 2) Then the parallel pool starts. I removed the reverse strand predictions block of code. The reverse strand predictions use the same function as forward strand, but sets the specific flag to true. This helps streamline the code for publication IMO. 3) A bedgraph format data frame is returned from the parallel pool worker. This will allow the workers to run asynchronously since the order no longer matters. 4) The bedgraph is sorted and written to a single bigwig file. I updated the parser to remove the s from the models flag. I also update prediction_tools.py with a function to make sure that all chromosomes are sorted in a bigwig friendly order (that means each chrom is ordered from 1-22,x,y). I updated the prediction data generator to implement single batch prediction in case the number of examples is less than the batch size which was causing an error with BED specific prediction.