#80 Update readme.md, parser.py, and create data.py
I updated the readme docs to include information on the hg38 reference genome. I added an optional subcommand `maxatac data` that will download the specific reference data to `~/opt/maxatac/data`. This is an optional approach to getting the data that we need downloaded. I updated the parser to include the subparser for `maxatac data`. I also removed the maxATAC_data submodule from the repo since we will have users download it on their own. The default reference genome for maxATAC is hg38. I edited the files to reference the dir downloaded by maxatac data.  The user still has the ability to supply any blacklist bed/bw file to the subcommands that use them.