#81 Update data.py moved package checks to system_tools.py. Updated help for subcommands.
Updated some of the paths to make sure that the opt directory exists. I tested on BMI and there was no directory at the user directory location of ~/opt so I made sure to create directory if these do not exist. They user can change the location of the data directory, but if they want to use the hardcoded paths (I know, it is awful, just a hack for now) they have to make sure to line up the constants.py file with the data directory they chose.