#93 Update the ROI import function for custom prediction regions.
Updated the code so that it only imports the first three columns of the ROI bed file and uses the prediction_tools.py function for importing prediction ROIs. The prediction regions method will merge intervals if they are close together. It will then window those regions using pybedtools to create sliding windows along the ROI. This still needs to be tested fully. This has been tested with data from @hrayjones mentioned in the comments. I also updated the readme information.