tests for average, prepare, and normalize
This commit has updates to the docs for prepare, average, and normalization. There have also been updated tests for these functions. They have been tested with data that will be uploaded to Zenodo. The parser.py file was also updated for those three functions to make them consistent across function. This might conflict with work that Faiz has done on the parse so we might need to merge conflicts later.